refactor(realm): extract usb bridge + device shared helpers from js-realm-shared
