+++ packages/webapp/src/kernel/realm/realm-usb-bridge.ts
@@ -1,0 +1,70 @@
+/**
+ * `realm-usb-bridge.ts` — the realm `usb` global mirroring WebUSB.
+ * Extracted from `js-realm-shared.ts`; no behavior change.
+ */
+import type { UsbControlSetup, UsbDeviceFilter, UsbDeviceInfo } from '../usb-device-registry.js';
+import {
+  asFilterArray,
+  bytesToDataView,
+  type DeviceRpc,
+  toRealmBytes,
+  type WireInResult,
+  type WireOutResult,
+} from './realm-device-shared.js';
+
+/** A realm-facing WebUSB device. Methods carry the opaque handle. */
+export interface RealmUsbDevice extends UsbDeviceInfo {
+  open(): Promise<void>;
+  close(): Promise<void>;
+  reset(): Promise<void>;
+  selectConfiguration(value: number): Promise<void>;
+  claimInterface(interfaceNumber: number): Promise<void>;
+  releaseInterface(interfaceNumber: number): Promise<void>;
+  controlTransferIn(
+    setup: UsbControlSetup,
+    length: number
+  ): Promise<{ status: string; data: DataView }>;
+  controlTransferOut(
+    setup: UsbControlSetup,
+    data: ArrayBuffer | ArrayBufferView
+  ): Promise<WireOutResult>;
+  transferIn(endpointNumber: number, length: number): Promise<{ status: string; data: DataView }>;
+  transferOut(endpointNumber: number, data: ArrayBuffer | ArrayBufferView): Promise<WireOutResult>;
+}
+
+export interface RealmUsbApi {
+  list(): Promise<RealmUsbDevice[]>;
+  request(filters?: UsbDeviceFilter | UsbDeviceFilter[]): Promise<RealmUsbDevice>;
+}
+
+function makeUsbDevice(rpc: DeviceRpc, info: UsbDeviceInfo): RealmUsbDevice {
+  const h = info.handle;
+  const toData = (r: WireInResult) => ({ status: r.status, data: bytesToDataView(r.bytes) });
+  return {
+    ...info,
+    open: () => rpc.call<void>('usb', 'open', [h]),
+    close: () => rpc.call<void>('usb', 'close', [h]),
+    reset: () => rpc.call<void>('usb', 'reset', [h]),
+    selectConfiguration: (value) => rpc.call<void>('usb', 'selectConfig', [h, value]),
+    claimInterface: (n) => rpc.call<void>('usb', 'claim', [h, n]),
+    releaseInterface: (n) => rpc.call<void>('usb', 'release', [h, n]),
+    controlTransferIn: async (setup, length) =>
+      toData(await rpc.call<WireInResult>('usb', 'controlIn', [h, setup, length])),
+    controlTransferOut: (setup, data) =>
+      rpc.call<WireOutResult>('usb', 'controlOut', [h, setup, toRealmBytes(data)]),
+    transferIn: async (ep, length) =>
+      toData(await rpc.call<WireInResult>('usb', 'transferIn', [h, ep, length])),
+    transferOut: (ep, data) =>
+      rpc.call<WireOutResult>('usb', 'transferOut', [h, ep, toRealmBytes(data)]),
+  };
+}
+
+/** Build the realm `usb` global. Exported for parity / unit tests. */
+export function createUsbBridge(rpc: DeviceRpc): RealmUsbApi {
+  return {
+    list: async () =>
+      (await rpc.call<UsbDeviceInfo[]>('usb', 'list', [])).map((i) => makeUsbDevice(rpc, i)),
+    request: async (filters) =>
+      makeUsbDevice(rpc, await rpc.call<UsbDeviceInfo>('usb', 'request', [asFilterArray(filters)])),
+  };
+}
