+++ packages/webapp/tests/kernel/realm/usb-bridge.test.ts
@@ -9,7 +9,7 @@
 
 import type { CommandContext } from 'just-bash';
 import { describe, expect, it } from 'vitest';
-import { createUsbBridge } from '../../../src/kernel/realm/js-realm-shared.js';
+import { createUsbBridge } from '../../../src/kernel/realm/realm-usb-bridge.js';
 import { attachRealmHost } from '../../../src/kernel/realm/realm-host.js';
 import { RealmRpcClient } from '../../../src/kernel/realm/realm-rpc.js';
 import type { UsbBackend } from '../../../src/shell/supplemental-commands/usb-backends.js';
